reafactory: refactory old functions
