+++ srcs/Server/Route.cpp
@@ -284,11 +284,3 @@ std::ostream &operator<<(std::ostream &os, Route const &route) {
 	return (os);
 }
 
-std::ostream &operator<<(std::ostream &os, RouteResponse const &route) {
-    os << std::endl;
-    os << "\tStatus Code: " << route.StatusCode << std::endl;
-    os << "\tFile Extension: " << route.FileExtension << std::endl;
-    if (route.RedirectPath != "")
-        os << "\tredirect path: " << route.RedirectPath << std::endl;
-	return (os);
-}
